+++ lib/rdkafka/admin/incremental_alter_configs_report.rb
@@ -2,7 +2,7 @@
 
 module Rdkafka
   class Admin
-    class DescribeConfigsReport
+    class IncrementalAlterConfigsReport
       attr_reader :resources
 
       def initialize(config_entries:, entry_count:)